+++ src/main/org/firebirdsql/ds/FBSimpleDataSource.java
@@ -451,6 +451,16 @@ public class FBSimpleDataSource extends RootCommonDataSource implements DataSour
         mcf.setIgnoreProcedureType(ignoreProcedureType);
     }
 
+    @Override
+    public boolean isWireCompression() {
+        return mcf.isWireCompression();
+    }
+
+    @Override
+    public void setWireCompression(boolean wireCompression) {
+        mcf.setWireCompression(wireCompression);
+    }
+
     /*
      * INTERFACES IMPLEMENTATION
      */