+++ src/main/org/firebirdsql/gds/ng/FbConnectionProperties.java
@@ -247,6 +247,9 @@ public final class FbConnectionProperties extends AbstractAttachProperties<IConn
             case isc_dpb_auth_plugin_list:
                 setAuthPlugins(parameter.getValueAsString());
                 break;
+            case isc_dpb_wire_compression:
+                setWireCompression(true);
+                break;
             case isc_dpb_utf8_filename:
                 // Filter out, handled explicitly in protocol implementation
                 break;