+++ src/main/org/firebirdsql/gds/ng/wire/version16/Version16Descriptor.java
@@ -50,6 +50,7 @@ public final class Version16Descriptor extends AbstractProtocolDescriptor implem
                 WireProtocolConstants.arch_generic,
                 WireProtocolConstants.ptype_lazy_send, // Protocol implementation expects lazy send
                 WireProtocolConstants.ptype_lazy_send,
+                true,
                 5);
     }