+++ src/main/org/firebirdsql/jdbc/FBConnectionProperties.java
@@ -67,6 +67,7 @@ public class FBConnectionProperties implements FirebirdConnectionProperties, Ser
     public static final String DATA_TYPE_BIND = "dataTypeBind";
     public static final String SESSION_TIME_ZONE = "sessionTimeZone";
     public static final String IGNORE_PROCEDURE_TYPE = "ignoreProcedureType";
+    public static final String WIRE_COMPRESSION = "wireCompression";
 
     private Map<String, Object> properties = new HashMap<>();
     private String type;
@@ -423,6 +424,16 @@ public class FBConnectionProperties implements FirebirdConnectionProperties, Ser
     @Override
     public void setIgnoreProcedureType(boolean ignoreProcedureType) {
         setBooleanProperty(IGNORE_PROCEDURE_TYPE, ignoreProcedureType);
+    }
+
+    @Override
+    public boolean isWireCompression() {
+        return getBooleanProperty(WIRE_COMPRESSION);
+    }
+
+    @Override
+    public void setWireCompression(boolean wireCompression) {
+        setBooleanProperty(WIRE_COMPRESSION, wireCompression);
     }
 
     public void setNonStandardProperty(String propertyMapping) {