+++ src/main/org/firebirdsql/jdbc/FirebirdConnectionProperties.java
@@ -542,4 +542,30 @@ public interface FirebirdConnectionProperties {
      * @since 3.0.6
      */
     void setIgnoreProcedureType(boolean ignoreProcedureType);
+
+    /**
+     * Get if wire compression should be enabled.
+     * <p>
+     * Wire compression requires Firebird 3 or higher, and the server must have the zlib library. If compression cannot
+     * be negotiated, the connection will be made without wire compression.
+     * </p>
+     * <p>
+     * This property will be ignored for native connections. For native connections, the configuration in
+     * {@code firebird.conf} read by the client library will be used.
+     * </p>
+     *
+     * @return {@code true} wire compression enabled
+     * @since 4.0
+     */
+    boolean isWireCompression();
+
+    /**
+     * Sets if the connection should try to enable wire compression.
+     *
+     * @param wireCompression
+     *         {@code true} enable wire compression, {@code false} disable wire compression (the default)
+     * @see #isWireCompression()
+     * @since 4.0
+     */
+    void setWireCompression(boolean wireCompression);
 }